flask-server/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.2.0
Werkzeug==2.3.7
openai==0.28.1
python-dotenv==1.0.0
google-generativeai==0.3.2
